restructuring + documentation
