+++ src/main/scala/org/kframework/minikore/parser/MiniToTextToMini.scala
@@ -3,7 +3,7 @@ package org.kframework.minikore.parser
 import org.apache.commons.io.FileUtils
 import org.kframework.minikore.implementation.DefaultBuilders
 import org.kframework.minikore.implementation.MiniKore.Definition
-import org.kframework.minikore.interfaces.BuilderInterface.Builders
+import org.kframework.minikore.interfaces.build.Builders
 
 // TODO(Daejun): drop this file
